Refactor WebSocket client and data handler
- Split the DataHandler into its own module for better organization.
- Update WebSocketClient to use the new DataHandler structure.
- Enhance WebSocketClient with improved connection handling, including reconnect logic and message processing.
- Implement detailed message handling for various TradingView events.
- Introduce a cancellation token for managing WebSocket closure.
- Refactor message sending logic to avoid deadlocks and improve clarity.
- Update error handling to provide more informative logging.
